requirements.txt
fastapi
uvicorn
openai
supabase
python-multipart
requests
python-dotenv